pyproject.toml
[project]
name = "accounts"
version = "0.1.0"
description = "FastAPI microservice for authentication and account management."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "fastapi-storages>=0.3.0",
    "fastapi[standard]>=0.119.1",
    "pillow>=12.0.0",
    "psycopg[binary]>=3.2.11",
    "pwdlib[argon2]>=0.2.1",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.27",
]

[tool.poe.tasks]
dev = "uvicorn app.main:app --reload"
alembic-upgrade = "alembic upgrade head"
alembic-revision = "alembic revision --autogenerate"

[tool.pyright]
reportExplicitAny = "none"
reportAny = "none"
reportUnknownVariableType = "none"
reportUnknownMemberType = "none"
reportMissingTypeStubs = "none"
reportUnusedCallResult = "none"
reportUnknownArgumentType = "none"
